+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [BREAKING] Move `get_assembly_op` to the `MastForest`, remove trait `MastNodeErrorContext` ([#2430](https://github.com/0xMiden/miden-vm/pull/2430)).
 - Run batch inversion concurrently per fragment in parallel trace generation ([#2405](https://github.com/0xMiden/miden-vm/issues/2405)).
 - Adds a cached commitment to the `MastForest` ([#2447](https://github.com/0xMiden/miden-vm/pull/2447))
+- Add MastForest validation ([#2412](https://github.com/0xMiden/miden-vm/pull/2412)).
 
 ## 0.20.0 (2025-12-05)
 
@@ -53,6 +54,7 @@
 - [BREAKING] Normalized naming of `verify` procedures of ECDSA precompile ([#2413](https://github.com/0xMiden/miden-vm/issues/2413)).
 - Refactored Blake3_256 fingerprints to allocate less ([#2375](https://github.com/0xMiden/miden-vm/pull/2375)).
 - [BREAKING] Normalized signature encoding methods in the `dsa` module of the core library.
+
 
 ## 0.19.1 (2025-11-6)